[Bugfix] : gameSetNo List 제외
