+++ src/main/java/com/server/gummymurderer/controller/GameNpcCustomController.java
@@ -13,8 +13,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-
 @RestController
 @RequestMapping("/api/v1/npc/custom")
 @RequiredArgsConstructor
@@ -23,10 +21,10 @@ public class GameNpcCustomController {
     private final GameNpcCustomService gameNpcCustomService;
 
     @PostMapping("/save")
-    public Response<GameNpcCustomSaveResponse> saveNpcCustom(@AuthenticationPrincipal CustomUserDetails customUserDetails, @RequestBody List<GameNpcCustomSaveRequest> requests) {
+    public Response<GameNpcCustomSaveResponse> saveNpcCustom(@AuthenticationPrincipal CustomUserDetails customUserDetails, @RequestBody GameNpcCustomSaveRequest request) {
 
         Member loginMember = customUserDetails.getMember();
-        GameNpcCustomSaveResponse response = gameNpcCustomService.npcCustomSave(loginMember, requests);
+        GameNpcCustomSaveResponse response = gameNpcCustomService.npcCustomSave(loginMember, request);
 
         return Response.success(response);
     }
