+++ src/main/java/com/server/gummymurderer/domain/dto/gameNpcCustom/GameNpcCustomSaveRequest.java
@@ -1,9 +1,8 @@
 package com.server.gummymurderer.domain.dto.gameNpcCustom;
 
-import com.server.gummymurderer.domain.entity.GameNpc;
-import com.server.gummymurderer.domain.entity.GameNpcCustom;
-import com.server.gummymurderer.domain.entity.GameSet;
 import lombok.*;
+
+import java.util.List;
 
 @Getter
 @Setter
@@ -13,22 +12,5 @@ import lombok.*;
 public class GameNpcCustomSaveRequest {
 
     private Long gameSetNo;
-    private String npcName;
-    private int mouth;
-    private int ear;
-    private int body;
-    private int tail;
-
-    public GameNpcCustom toEntity(GameSet gameSet, GameNpc gameNpc) {
-
-        return GameNpcCustom.builder()
-                .gameSet(gameSet)
-                .gameNpc(gameNpc)
-                .mouth(mouth)
-                .ear(ear)
-                .body(body)
-                .tail(tail)
-                .build();
-
-    }
+    private List<NpcCustomInfo> npcCustomInfos;
 }
